Add an OAuth callback route for after the OAuth login flow finishes.
The OAuth flow is triggered automatically as soon as the
openshiftResourceParser component is used, and there isn't an access
token yet.
If a token already exists, the login flow is skipped and the requested
OpenShift API call is made.

After the login flow completes, the browser is redirected to the
original page they were on when the login flow was started.