Allow images to be bundled within the walkthroughs themselves
Currently images are included in a separate images directory which
is provided by the server. However, when users want to add their
own walkthroughs they won't be able to add images directly to this
repo. It'd also mean that their adoc wouldn't render correctly
locally as the image would be in a separate location.

A second issue also resolved by this commit is that, although you can
specify WALKTHROUGH_LOCATIONS it does not reflect in the web app.

The changes made in this commit provide a dynamic static route in
the server to allow for the files of each walkthrough to be served
using the same path as if you were previewing them locally. For
example: image::my.png will now work both locally and in the web app
(assuming that my.png is in the same directory as the adoc file).

The frontend has been updated to accomodate for these changes. Using
the new endpoint instead of the static /public endpoints from before.

Finally, the images from /public/images have been moved into their
respective walkthroughs and placed in an images/ subdirectory.