Allow branches and tags to be specified in a git clone
Currently during a git clone, we only allow cloning the default
branch of a repo. However, it may be useful for releases to be
able to have branches and tags specified in the URL.

This change follows the syntax of npm repository dependencies, using
the hash value of the URL to specify the branch or tag. If none are
specified it will clone the default branch of the repo.