INTLY-107 Allow single step tasks without a title
Currently, when creating a Walkthrough if a writer would like to
have a task that contains a single step they must still define a
h3 block and give the block a title. Otherwise features like the
verification block cannot be used.

In some cases, when creating this type of single step task, the
title of the task is the title of the step also.

This change allows tasks to contain verification blocks, instead
of them only being allowed in steps. Meaning that a writer can
create a single step task without having to specify a h3 block.

Verification blocks will still work as expected in the web app.

Tasks like below should now work and render as expected:

```
== My Task

Below is a verification block, it will render even though it's
outside of a Step.

[type=verification]
This is a task-level verification block
```