INTLY-380 Use React Router instead of window.href for logo click
Currently, when the "Red Hat Solution Explorer" logo in the top
left-hand corner is clicked we set window.href explicitly, this
can result in a full refresh of the page, which then results in
the loading screen for the web app revealing itself momentarily.

We do not need to perform this kind of full refresh normally.
Instead we can use the react router to change the state of the app
so that it's on the landing page, this results in a much faster
transition without a loading screen.

Verification:
* On the landing page, select "Red Hat Solution Explorer"
* The page should appear again immediately
* Do the same from a Walkthrough overview page to ensure the loading
screen doesn't appear.