INTLY-399 Provide all attributes to overview asciidoc parser
Currently, on the overview screen for a walkthrough, we do not
provide all of the attributes that we provide to each step. This
seems to have been an oversight from the pre-custom-walkthrough
format.

This change provides all of the attribtues that we see in each
walkthrough step to the overview page so that things like resources
on the right-hand side can work.