INTLY-418 Use new loading screen
The current loading screen being used does not display enough
information about the services being provisioned, the percentage
usually stalls at a percentage as one service takes much longer to
provision than others. This leads the end-users of the web app to
think that something has gone wrong.

The new loading screen has a per-service status view which is a
better indicator of what's going on in the background.