INTLY-418 Add functions for distinguishing provisioning states
Currently the ability to distinguish between various provisioning
states isn't self-contained. This change moves the logic for
determining the provisioning state of a service instance to a module
which can be reused. It narrows everything down to three states:

* Provisioning
* Provisioned
* Error

The UI has also been updated to reflect these states with the error
message being shown when it is available.

Verification:
* Cause a broken ServiceInstance either by incorrect plan name or
other means.
* Open the web app.
* It should start provisioning the services.

There's also a helper function for retrieving the error of a
service instance