INTLY-398 Restore task progress when coming from overview screen
Currently, in a situation where a user enters a Walkthrough through
the overview screen they are brought to task 0 even if they have
made progress.

This change ensures that a user will be brought to the last step
that the user checked a verification box on. This is the same
functionality as entering a Walkthrough from the landing page.

There are also changes included which resolve various minor issues
noticed. Mostly around values being undefined.

Verification:
- Go into a walkthrough with no previous progress
- Click get started on the overview screen
- Ensure you are taken to task 0
- Complete a couple of tasks and take note of the task you last
verified.
- Click the name of the walkthrough in the breadcrumbs
- On the overview screen, click get started
- Ensure you are taken back to the task