INTLY-185 Skip importing walkthrough if walkthrough.json is absent
JIRA: https://issues.jboss.org/browse/INTLY-185

Currently, when walkthrough.json does not exist in a walkthrough
directory it is still shown in the landing page. This can result
in issues when starting/going through a walkthrough.

This change filters out any walkthroughs that do not include a
walkthrough.json and a walkthrough.adoc file. If these are not
included then a message is logged to the console about the
walkthrough not being included, so that the user knows what is
going wrong.

Verification:
- Remove walkthrough.json from any walkthrough
- Restart the server
- Ensure the walkthrough is not included on the landing page
- Ensure there is a log from the server about the walkthrough
being skipped