[WIP] - PF4 update with master (#382)
* INTLY-398 Restore task progress when coming from overview screen

Currently, in a situation where a user enters a Walkthrough through
the overview screen they are brought to task 0 even if they have
made progress.

This change ensures that a user will be brought to the last step
that the user checked a verification box on. This is the same
functionality as entering a Walkthrough from the landing page.

There are also changes included which resolve various minor issues
noticed. Mostly around values being undefined.

Verification:
- Go into a walkthrough with no previous progress
- Click get started on the overview screen
- Ensure you are taken to task 0
- Complete a couple of tasks and take note of the task you last
verified.
- Click the name of the walkthrough in the breadcrumbs
- On the overview screen, click get started
- Ensure you are taken back to the task

* fix(INTLY-525): fix content wrapping issues and alignments

* INTLY-383 If preamble doesn't have lines ignore shortDescription (#371)

JIRA: https://issues.jboss.org/browse/INTLY-383

Currently, in the server-side parsing of the walkthrough files, if
the walkthrough has a preamble that does not start with a paragraph
containing at least one line then the parsing will fail. This is
due to the parser expecting that a walkthrough preamble being a
paragraph.

This change performs a check on the preamble to ensure that it
checks before trying to retrieve the shortDescription that the
preamble starts with a paragraph that contains at least one line.

Verification:
- Put a bulleted list as the preamble of a paragraph (first content
after the title)
- Restart the server and ensure it does not error
- Go through the walkthrough, ensure it works and the overview
renders correctly

* Add release steps for webapp & walkthroughs

* INTLY-379 Stop loading screen between tasks in walkthrough

Currently when a user selects to go to the next task in a
walkthrough a loading screen will appear momentarily. This is due
to the request for the asciidoc contents and manifest of the
walkthrough being requested each time the task component mounts.
This request clears the current Redux store, causing the loading
page to appear until the store is populated again.

However, when moving between tasks in the same walkthrough this
does not need to be done as long as the asciidoc content and the
manifest have already been retrieved and stored in Redux.

This change performs a check before trying to kick off the
requests to retrieve this information. The check performed when
the task component mounts is:
- Is this task for the same walkthrough as the last task
- Is the asciidoc content already available in the Redux store
- Is the manifest content already available in the Redux store

If these are true, the loading screen does not need to be shown.

Verification:
- Go into any walkthrough with more than one task
- Click through the tasks of the walkthrough, ensure the loading
screen is not shown
- Click on the 'home' icon in the breadcrumbs to go to the landing
page
- Click on any other walkthrough and ensure the loading screen is
shown (it's a new walkthrough, we need to load again)
- Perform a hard refresh on a task, ensure that a loading screen
is shown (we have no knowledge of the last walkthrough)

* INTLY-185 Skip importing walkthrough if walkthrough.json is absent

JIRA: https://issues.jboss.org/browse/INTLY-185

Currently, when walkthrough.json does not exist in a walkthrough
directory it is still shown in the landing page. This can result
in issues when starting/going through a walkthrough.

This change filters out any walkthroughs that do not include a
walkthrough.json and a walkthrough.adoc file. If these are not
included then a message is logged to the console about the
walkthrough not being included, so that the user knows what is
going wrong.

Verification:
- Remove walkthrough.json from any walkthrough
- Restart the server
- Ensure the walkthrough is not included on the landing page
- Ensure there is a log from the server about the walkthrough
being skipped

* fix(content): Fix Note blocks to render correctly

* feat(PF4): Introduce PatternFly 4

* fix(tests): fix failing eslint tests

* fix(dependencies): revert to correct version of PF React-Core

* fix(dependencies): Update CSS compiler dependencies

* fix(breadcrumbs): revert to PF3 breadcrumbs

* fix(lists): remove indentation from walkthrough resources list

* fix(tests): fix broken tests

* fix(css): code cleanup and postcss fixes

* feat(PF4): Introduce PatternFly 4

* fix(lists): remove indentation from walkthrough resources list

* fix(INTLY-525): fix content wrapping issues and alignments

* Update PF4 with latest Master

* merge master into PF4

* Update packages for PF4 branch

* fix merge issues