Update about endpoint and add per-walkthrough about endpoint
Currently we have an `/about` endpoint, but this excludes any
local path walkthroughs that are provided via WALKTHROUGH_LOCATIONS
and does not include a lot of information about the location.

It'd also be nice to be able to retrieve the info for the
WALKTHROUGH_LOCATION that any walkthrough belongs to. This change
also adds a `/about/walkthrough/:walkthroughId` endpoint, where
the ID of a walkthrough can be provided and the location which it's
from will be returned.

Verification:
- Start the server with `yarn start:dev`
- `curl http://localhost:3006/about`
- Ensure the response contains required info for the WALKTHROUGH_LOCATIONS
- Pick a walkthrough from the UI, get it's ID from the address bar
- `curl http://localhost:3006/about/<walkthrough-id>`
- Ensure the correct WALKTHROUGH_LOCATION info has been returned and
contains the required information