INTLY-1241 Allow for a new env var for the threescale wildcard domain
If THREESCALE_WILDCARD_DOMAIN is set, it will be made available in the
ui as window.OPENSHIFT_CONFIG.threescaleWildcardDomain.

This in turn will be used to set the openshift-app-host property for
walkthroughs to indicate the wildcard suffix/domain, which is useful in
a cluster where 3scale has the wildcard proxy setup.
This is used in walkthrough 2 for the staging base url.