INTLY-1362 Ensure AMQ Online shows up in the installed apps view
Currently, when AMQ Online is provisioned as part of Walkthrough 1A
it will be visible in the installed apps view until the user
refreshes. This is because the websocket watch for AMQ Online is
only started when Walkthrough 1A is started.

This change does a small refactor to provide two lists about services,
WATCH_SERVICES and PROVISION_SERVICES.
- WATCH_SERVICES is a list of service instance names that should be
watched when the web app starts up.
- PROVISION_SERVICES is a list of service instance names that should
be provisioned when the web app starts up.

These two lists can now change independently, meaning we can tell the
web app to look for AMQ Online but not provision it when it starts up.

An issue that came out of this is that services in the installed apps
view may not have a completely set order. Because of this when a user
navigates to the landing page after completing Walkthrough 1A they'll
see AMQ Online at the bottom of the list. However, when they refresh
it'll be near the top.

A change has been added to allow these services to be sorted before
rendering the list. Meaning the list should be consistent in all
cases.

Finally, a patch has been made to remove appending product details
to the ServiceInstance resource object we retrieve from OpenShift.
Instead these details will be retrieved when required through a
`getProductDetails` function.