INTLY-1362 Allow AMQ Online to be launched from the landing page
Currently AMQ Online is only provisioned when going into a walkthrough
that depends on a slice of it in their JSON configuration. However,
users may want to experience AMQ Online without starting a walkthrough
and they may not even know AMQ Online is available if it isn't shown
on the landing page right hand side at first.

This commit makes a change to allow for a set of services to display
to be defined. These services will always be shown in the UI on the
landing page, however they will not necessarily be provisioned when
the web app starts (this is defined in a different list).

If the service is not provisioned there will be a launch button next
to it in the UI. As each service requires different steps to be
performed to provision it, only AMQ Online has been included in this
change. However the pattern has been set up to allow for others.