INTLY-1362 Allow display services to be configured from the server
Currently, we don't allow services that should be shown even if
they're not yet provisioned to be configured by the end-user. This
may result in backwards-compatability issues as there are users who
do not want to provision every service or have it visible.

This change allows the InstallAppsView component to take in a list
of information provided in the custom server config that specifies
which services to display. This will override the default option.