use managed fuse for walkthrough 1A
Use the managed fuse provisioned by the integreatly operator for
walkthrough 1A instead of provisioning Fuse in the user namespace.
By doing this, we won't need to add the pull secret to the walkthrough
namespaces and no extra permissions are needed to be provided for the dev
user to create subscriptions.