fix state caching between walkthroughs
currently, if a user starts a walkthrough and then, without
performing a refresh of their page, navigates to another walkthrough,
it causes an issue with the manifest of the previous walkthrough
being used to provision services in the new walkthrough.

this results in services not being provisioned when they are
required in walkthroughs.

this change ensures that navigating back to the main page will
ensure the manifest cache for the previous walkthrough is cleared.

verification:
- have an existing rhmi installation from master
- update the web app image to point to quay.io/aidenkeating/tutorial-web-app:fix-caching-2
- open walkthrough 2 (3scale) to the first page, no services should be provisioned
- without refreshing, navigate to walkthrough 1, ensure that fuse and amq are provisioned